dom/origin-trials/tests/mochitest/test_trial_hidden.html

<!doctype html>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="common.js"></script>
<script>
  assertTestTrialActive(false);
  add_task(function() {
    ok(!navigator.privateAttribution, "PrivateAttribution trial disabled.");
    ok(!('PrivateAttribution' in window), "PrivateAttribution trial disabled.");
  });
</script>
